refactor all

+++ src/main/java/com/example/justatest/post/PostService.java
@@ -2,6 +2,7 @@ package com.example.justatest.post;
 
 import com.example.justatest.dto.PostRequestDto;
 import com.example.justatest.dto.PostResponseDto;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
@@ -22,9 +23,10 @@ public class PostService {
                 .toList();
     }
 
-    public Optional<PostResponseDto> findById(UUID uuid) {
+    public PostResponseDto findById(UUID uuid) {
         return postRepository.findById(uuid)
-                .map(post -> modelMapper.map(post, PostResponseDto.class));
+                .map(post -> modelMapper.map(post, PostResponseDto.class))
+                .orElseThrow(EntityNotFoundException::new);
     }
 
     public PostResponseDto save(PostRequestDto postRequestDto) {
@@ -32,20 +34,24 @@ public class PostService {
         return modelMapper.map(postRepository.save(post), PostResponseDto.class);
     }
 
-    public Optional<PostResponseDto> update(UUID uuid, PostRequestDto postRequestDto) {
+    public PostResponseDto update(UUID uuid, PostRequestDto postRequestDto) {
         Optional<Post> postOptional = postRepository.findById(uuid);
 
         if (postOptional.isEmpty()) {
-            return Optional.empty();
+            throw new EntityNotFoundException();
         }
 
         Post post = postOptional.get();
         modelMapper.map(postRequestDto, post);
 
-        return Optional.of(modelMapper.map(postRepository.save(post), PostResponseDto.class));
+        return modelMapper.map(postRepository.save(post), PostResponseDto.class);
     }
 
     public void deleteById(UUID uuid) {
+        if (!postRepository.existsById(uuid)) {
+            throw new EntityNotFoundException();
+        }
+
         postRepository.deleteById(uuid);
     }
 }
